+++ lib/src/index.ts
@@ -1,14 +1,4 @@
 import 'reflect-metadata';
 
 export * from './json';
-
-// Internals (for backend implementations)
-export * from './metadata';
-export * from './types';
-export * from './options';
-export { default as FrontendFunctions } from './frontend_functions';
-export { default as DecoratorFactory } from './decorator_factory';
-export { default as ObjectPropertySerializer } from './object_property_serializer';
-export { default as PropertySerializer } from './property_serializer';
-export { default as TypeSerializerPicker } from './type_serializer_picker';
-export { default as TypeSerializer } from './type_serializer';
+export { Serializable } from './types';